Choose colors by hash
