+++ libs/ngx-charts-on-fhir/src/lib/data-layer/data-layer-manager.service.ts
@@ -157,7 +157,6 @@ export class DataLayerManagerService {
   reset() {
     this.cancel$.next();
     this.state = { ...this.state, layers: {}, selected: [] };
-    this.colorService.reset();
   }
 
   /**
