+++ libs/ngx-charts-on-fhir/src/lib/utils.ts
@@ -86,3 +86,13 @@ export function formatMonthRange(range: MonthRange) {
     return formatDays(range);
   }
 }
+
+export function hashCode(str: string): number {
+  let hash = 0;
+  for (let i = 0, len = str.length; i < len; i++) {
+    let chr = str.charCodeAt(i);
+    hash = (hash << 5) - hash + chr;
+    hash |= 0; // Convert to 32bit integer
+  }
+  return hash;
+}
